モデル生成関数を切り出し
